Extracted updateOneRecordMutation
