+++ packages/twenty-front/src/modules/object-record/hooks/useUpdateOneRecord.ts
@@ -6,6 +6,7 @@ import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadat
 import { getRecordNodeFromRecord } from '@/object-record/cache/utils/getRecordNodeFromRecord';
 import { updateRecordFromCache } from '@/object-record/cache/utils/updateRecordFromCache';
 import { getUpdateOneRecordMutationResponseField } from '@/object-record/hooks/useGenerateUpdateOneRecordMutation';
+import { useUpdateOneRecordMutation } from '@/object-record/hooks/useUpdateOneRecordMutation';
 import { ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { sanitizeRecordInput } from '@/object-record/utils/sanitizeRecordInput';
 
@@ -24,8 +25,15 @@ export const useUpdateOneRecord = <
 }: useUpdateOneRecordProps) => {
   const apolloClient = useApolloClient();
 
-  const { objectMetadataItem, updateOneRecordMutation, getRecordFromCache } =
-    useObjectMetadataItem({ objectNameSingular }, depth, queryFields);
+  const { objectMetadataItem, getRecordFromCache } = useObjectMetadataItem(
+    { objectNameSingular },
+    depth,
+    queryFields,
+  );
+
+  const { updateOneRecordMutation } = useUpdateOneRecordMutation({
+    objectNameSingular,
+  });
 
   const { objectMetadataItems } = useObjectMetadataItems();
