+++ packages/twenty-front/src/modules/object-record/hooks/useUpdateOneRecordMutation.ts
@@ -1,0 +1,23 @@
+import { useObjectMetadataItemOnly } from '@/object-metadata/hooks/useObjectMetadataItemOnly';
+import { useGenerateUpdateOneRecordMutation } from '@/object-record/hooks/useGenerateUpdateOneRecordMutation';
+
+export const useUpdateOneRecordMutation = ({
+  objectNameSingular,
+  computeReferences = false,
+}: {
+  objectNameSingular: string;
+  computeReferences?: boolean;
+}) => {
+  const { objectMetadataItem } = useObjectMetadataItemOnly({
+    objectNameSingular,
+  });
+
+  const updateOneRecordMutation = useGenerateUpdateOneRecordMutation({
+    objectMetadataItem,
+    computeReferences,
+  });
+
+  return {
+    updateOneRecordMutation,
+  };
+};
