+++ packages/twenty-front/src/modules/views/hooks/internal/usePersistViewFilterRecords.ts
@@ -8,21 +8,22 @@ import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadata
 import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadataItems';
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
 import { useCreateOneRecordMutation } from '@/object-record/hooks/useCreateOneRecordMutation';
+import { useUpdateOneRecordMutation } from '@/object-record/hooks/useUpdateOneRecordMutation';
 import { ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { GraphQLView } from '@/views/types/GraphQLView';
 import { ViewFilter } from '@/views/types/ViewFilter';
 
 export const usePersistViewFilterRecords = () => {
-  const {
-    updateOneRecordMutation,
-    deleteOneRecordMutation,
-    objectMetadataItem,
-    getRecordFromCache,
-  } = useObjectMetadataItem({
+  const { deleteOneRecordMutation, objectMetadataItem, getRecordFromCache } =
+    useObjectMetadataItem({
+      objectNameSingular: CoreObjectNameSingular.ViewFilter,
+    });
+
+  const { createOneRecordMutation } = useCreateOneRecordMutation({
     objectNameSingular: CoreObjectNameSingular.ViewFilter,
   });
 
-  const { createOneRecordMutation } = useCreateOneRecordMutation({
+  const { updateOneRecordMutation } = useUpdateOneRecordMutation({
     objectNameSingular: CoreObjectNameSingular.ViewFilter,
   });
