+++ packages/twenty-front/src/modules/views/hooks/internal/usePersistViewSortRecords.ts
@@ -8,21 +8,22 @@ import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadata
 import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadataItems';
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
 import { useCreateOneRecordMutation } from '@/object-record/hooks/useCreateOneRecordMutation';
+import { useUpdateOneRecordMutation } from '@/object-record/hooks/useUpdateOneRecordMutation';
 import { ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { GraphQLView } from '@/views/types/GraphQLView';
 import { ViewSort } from '@/views/types/ViewSort';
 
 export const usePersistViewSortRecords = () => {
-  const {
-    updateOneRecordMutation,
-    deleteOneRecordMutation,
-    objectMetadataItem,
-    getRecordFromCache,
-  } = useObjectMetadataItem({
+  const { deleteOneRecordMutation, objectMetadataItem, getRecordFromCache } =
+    useObjectMetadataItem({
+      objectNameSingular: CoreObjectNameSingular.ViewSort,
+    });
+
+  const { createOneRecordMutation } = useCreateOneRecordMutation({
     objectNameSingular: CoreObjectNameSingular.ViewSort,
   });
 
-  const { createOneRecordMutation } = useCreateOneRecordMutation({
+  const { updateOneRecordMutation } = useUpdateOneRecordMutation({
     objectNameSingular: CoreObjectNameSingular.ViewSort,
   });
